requirements.txt
ccxt>=4.4.0
Flask>=3.0.0
python-dotenv>=1.0.1
requests>=2.32.0
